# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ordnung",
    "description": "Local API of Ordnung, the private AI secretary for life admin.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/drafts/{draft_id}/proof": {
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "Get Proof",
        "description": "A letter's proof: tracking number, proofs with what each shows, timeline, what's missing and\nwhat the letter waits for.",
        "operationId": "get_proof_api_drafts__draft_id__proof_get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProofOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/tracking": {
      "put": {
        "tags": [
          "drafts"
        ],
        "summary": "Set Tracking",
        "description": "Save the tracking number of a sent letter (checked: a mistyped check digit is refused).",
        "operationId": "set_tracking_api_drafts__draft_id__tracking_put",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProofOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/proofs": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Add Proof",
        "description": "Attach a proof file to a sent letter. The file is kept private: it is never sent to AI.",
        "operationId": "add_proof_api_drafts__draft_id__proofs_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_proof_api_drafts__draft_id__proofs_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProofOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/proofs/{proof_id}": {
      "patch": {
        "tags": [
          "drafts"
        ],
        "summary": "Update Proof",
        "description": "Correct what a proof is, the day it shows or its note.",
        "operationId": "update_proof_api_drafts__draft_id__proofs__proof_id__patch",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          },
          {
            "name": "proof_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proof Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProofPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProofOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "drafts"
        ],
        "summary": "Remove Proof",
        "description": "Remove a proof; its file is deleted for good unless another proof uses it.",
        "operationId": "remove_proof_api_drafts__draft_id__proofs__proof_id__delete",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          },
          {
            "name": "proof_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proof Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProofOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/proof.pdf": {
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "Nachweis Pdf",
        "description": "The Nachweis: a summary with the timeline, the letter as sent and every proof file, as one PDF.",
        "operationId": "nachweis_pdf_api_drafts__draft_id__proof_pdf_get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/calls": {
      "get": {
        "tags": [
          "waiting"
        ],
        "summary": "List Calls",
        "description": "Call notes, newest call first (of one person or organisation, or one thread).",
        "operationId": "list_calls_api_calls_get",
        "parameters": [
          {
            "name": "party_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Party Id"
            }
          },
          {
            "name": "case_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CallNote"
                  },
                  "title": "Response List Calls Api Calls Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "waiting"
        ],
        "summary": "Create Call",
        "description": "Note a phone call. Nothing is sent anywhere; no AI reads it.",
        "operationId": "create_call_api_calls_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallNoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallNote"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/calls/{call_id}": {
      "patch": {
        "tags": [
          "waiting"
        ],
        "summary": "Update Call",
        "description": "Say the promise made on the call was kept (or take that back).",
        "operationId": "update_call_api_calls__call_id__patch",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallNotePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallNote"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "waiting"
        ],
        "summary": "Delete Call",
        "description": "Delete a call note.",
        "operationId": "delete_call_api_calls__call_id__delete",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_add_proof_api_drafts__draft_id__proofs_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "The proof: a photo or PDF of a receipt, a fax report, an e-mail …"
          },
          "kind": {
            "type": "string",
            "enum": [
              "posting_receipt",
              "delivery_record",
              "return_receipt",
              "fax_report",
              "sent_email",
              "cancel_confirmation",
              "other"
            ],
            "title": "Kind",
            "description": "What the proof is"
          },
          "on_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Date",
            "description": "The day it shows (posted, delivered …)"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "file",
          "kind"
        ],
        "title": "Body_add_proof_api_drafts__draft_id__proofs_post"
      },
...
      "CallNote": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "party_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Party Id"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "called_on": {
            "type": "string",
            "title": "Called On"
          },
          "contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "promise": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promise"
          },
          "promise_due": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promise Due"
          },
          "promise_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promise Amount"
          },
          "promise_kept_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promise Kept On"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "party_id",
          "case_id",
          "called_on",
          "contact",
          "summary",
          "promise",
          "promise_due",
          "promise_amount",
          "promise_kept_on",
          "created_at",
          "updated_at"
        ],
        "title": "CallNote",
        "description": "A phone call the person noted (Gesprächsnotiz): when, with whom, what was said and what was\npromised. A promise with a date is waited for (``secretary.waiting``)."
      },
      "CallNoteCreate": {
        "properties": {
          "party_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Party Id"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "called_on": {
            "type": "string",
            "title": "Called On"
          },
          "contact": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact"
          },
          "summary": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Summary"
          },
          "promise": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Promise"
          },
          "promise_due": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promise Due"
          },
          "promise_amount": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Promise Amount"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "called_on",
          "summary"
        ],
        "title": "CallNoteCreate",
        "description": "A phone call to note: when, with whom, what was said and what they promised (a promise with a\nday is waited for)."
      },
      "CallNotePatch": {
        "properties": {
          "kept": {
            "type": "boolean",
            "title": "Kept"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kept"
        ],
        "title": "CallNotePatch",
        "description": "Whether the call's promise was kept."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProofEntry": {
        "properties": {
          "proof": {
            "$ref": "#/components/schemas/Proof"
          },
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Document"
              },
              {
                "type": "null"
              }
            ]
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "shows": {
            "type": "string",
            "title": "Shows"
          },
          "does_not_show": {
            "type": "string",
            "title": "Does Not Show"
          }
        },
        "type": "object",
        "required": [
          "proof",
          "document",
          "label",
          "shows",
          "does_not_show"
        ],
        "title": "ProofEntry",
        "description": "A proof with its file and, in code-written words, what it shows and what it does not."
      },
...
      "ProofOverview": {
        "properties": {
          "draft_id": {
            "type": "string",
            "title": "Draft Id"
          },
          "sent": {
            "type": "boolean",
            "title": "Sent"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "tracking": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrackingInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "proofs": {
            "items": {
              "$ref": "#/components/schemas/ProofEntry"
            },
            "type": "array",
            "title": "Proofs"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/ProofEvent"
            },
            "type": "array",
            "title": "Timeline"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing"
          },
          "waiting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WaitingEntry"
              },
              {
                "type": "null"
              }
            ]
          },
          "caveat": {
            "type": "string",
            "title": "Caveat",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "draft_id",
          "sent",
          "channel",
          "tracking",
          "proofs",
          "timeline",
          "missing",
          "waiting",
          "caveat"
        ],
        "title": "ProofOverview",
        "description": "``GET /api/drafts/{id}/proof``: a letter's tracking number, proofs, timeline, what is missing and\nwhat it waits for."
      },
      "ProofPatch": {
        "properties": {
          "kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "posting_receipt",
                  "delivery_record",
                  "return_receipt",
                  "fax_report",
                  "sent_email",
                  "cancel_confirmation",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "on_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Date"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProofPatch",
        "description": "Corrections to a proof: what it is, the day it shows (``null`` removes it) and a note."
      },
...
      "TrackingInfo": {
        "properties": {
          "number": {
            "type": "string",
            "title": "Number"
          },
          "display": {
            "type": "string",
            "title": "Display"
          },
          "format": {
            "type": "string",
            "enum": [
              "s10",
              "domestic"
            ],
            "title": "Format"
          },
          "checked": {
            "type": "boolean",
            "title": "Checked"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "number",
          "display",
          "format",
          "checked",
          "note"
        ],
        "title": "TrackingInfo",
        "description": "A letter's tracking number as Ordnung read it (``drafts.proof.parse_tracking_number``)."
      },
      "TrackingUpdate": {
        "properties": {
          "tracking_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Number"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TrackingUpdate",
        "description": "A sent letter's tracking number (``null`` or empty removes it)."
      },
...
    }
  }
}